\name{tm.lexis.canada-package}
\alias{tm.lexis.canada-package}
\alias{tm.lexis.canada}
\docType{package}
\title{
tm.lexis.canada
}
\description{
Description: Databases of news articles often contain useful metadata which is formatted in different ways. The tm package in R is highly extendible in that it lends itself to producing plugin packages that read textual material in which is formatted in different ways.  This package reads in news articles from the Lexis-Nexis database for analysis in the tm package.  Note that there is additional useful metadata in the Lexis-Nexis documents (i.e. article titles), but it is not immediately clear that it is extractable in an automated way because there is no consistent tagging.

}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{tm.lexis.canada}
\packageIndices{tm.lexis.canada}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Simon J. Kiss, Wilfrid Laurier University

Maintainer: Simon J. Kiss < skiss at wlu dot ca>
}
\references{
Feinerer, I. “Introduction to the Tm Package Text Mining in R,” July 10, 2014.
}

\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
